refactor(orchestrator): extract ProbabilityUIMixin from the God-file (D-164 Slice 3)
Move the deterministic probability Generative-UI builders verbatim into a cohesive
ProbabilityUIMixin (SRP): concrete-label extraction (_extract_concrete_events +
_CONCRETE_EVENT_PATTERNS), _detect_probability_tree, _build_tree_structure,
_build_calculated_tree_props, _build_calculated_ui (~430 lines), _build_probability_tree_props,
_enrich_tree_labels_with_llm. _stream_socratic_evaluation stays in the client between the two
extracted regions, which are placed adjacent in the mixin so test_d123's
_build_calculated_ui..._build_probability_tree_props slice still bounds the method body.

OrchestratorClient(ProbabilityTutorBrain, StreamNormalizationMixin, TextStreamingMixin,
ProbabilityUIMixin) — brain stays first in the MRO. Append to TUTOR_SOURCE_FILES (one line);
convert test_d123 to read_tutor_source(); make the parity gate wrapping-tolerant (regex) since
the 4-base signature now wraps; fix the manifest parser to anchor on the assignment (was
matching the docstring mention). God-file 3,521 -> 2,786 lines (3,832 -> 2,786 across Slices 1-3).
